Split up API serializers
