+++ api/tests/serializers_tests.py
@@ -6,7 +6,7 @@ from rest_framework.authtoken.models import Token
 from rest_framework.test import APITestCase
 from waffle.models import Flag
 
-from api.serializers import FlagSerializer
+from api.serializers.privaterelay import FlagSerializer
 from emails.models import RelayAddress
 from emails.tests.models_tests import make_free_test_user, make_premium_test_user
